Testing:  Comments about the NeurCOMP class.
It's not clear (yet) how things ought to be structured; Prof. Bergel's design was 
different, and he built a single interface into his testing class in order to derive
Neurons, whereas I pushed such a thing into the initialization of the NeurCOMP
instance, but that forced me to add these step message sends in several places,
instead of having them in one place, tucked inside the 
  PerceptronTest1>>#newNeuron
method, which is how Prof. Bergel did it.