Chapter 3, section 4:  Error Backward Propagation.
Second step of back propagation consists of propagating errors
computed in the output layer back in the network.

This adds in a message for an NNetwork that tells the network
to do backward-propagation of errors through the network
given the expected outputs.

This is in method
  NNetwork>>#backwardPropagateError:

Added method NNetwork>>#outputLayer to support this change.