Chapter 3, section 4:  Error Backward Propagation.
Second step of back propagation consists of propagating errors
computed in the output layer back in the network.

This adds in the backward-propagation messages for the 
NeuronLayer instances which reside inside an NNetwork.
The backward-propagation process is started by sending
the NNetwork the backwardPropagateError: message and 
handing it the expectedOutputs.

This design allows for an arbitrary number of NNetwork 
NeuronLayers, as the layers have references to their
predcessors (via the previousLayer message/accessor.)

Also added is the 
  Neuron>>#adjustDeltaWith:
message which is sent by a NeuronLayer to its Neurons
during the backward-propagation process.