Chapter 5, Section 3: Visualizations of Learning, Network structure.
Helpful for visualizating neural net training/learning, and the network 
structure.

Also note that there are methods with pragmas used to indicate to the 
Roassal2 subsystem/framework that new pages need to be added when
visualizing in a Playground.  Very helpful example of how to use Roassal2
mechanisms to plug into what is already built into Pharo and Roassal2.